# perf(changelog): improve changelog sort performance